Add NSF Upgrade scenario definitions and improve interface validation logic.
