+++ tests/integration/system/nsf/BUILD.bazel
@@ -72,6 +72,7 @@ cc_library(
         "//tests/integration/system/nsf/interfaces:component_validator",
         "//tests/integration/system/nsf/interfaces:flow_programmer",
         "//tests/integration/system/nsf/interfaces:image_config_params",
+        "//tests/integration/system/nsf/interfaces:scenario",
         "//tests/integration/system/nsf/interfaces:test_params",
         "//tests/integration/system/nsf/interfaces:testbed",
         "//tests/integration/system/nsf/interfaces:traffic_helper",
