+++ tests/integration/system/nsf/interfaces/test_params.h
@@ -23,6 +23,7 @@
 #include "tests/integration/system/nsf/interfaces/component_validator.h"
 #include "tests/integration/system/nsf/interfaces/flow_programmer.h"
 #include "tests/integration/system/nsf/interfaces/image_config_params.h"
+#include "tests/integration/system/nsf/interfaces/scenario.h"
 #include "tests/integration/system/nsf/interfaces/testbed.h"
 #include "tests/integration/system/nsf/interfaces/traffic_helper.h"
 #include "thinkit/ssh_client.h"
@@ -35,7 +36,6 @@ namespace pins_test {
 // parameterized NSF integration test.
 struct NsfTestParams {
   std::string name;
-  std::string test_case_id;
   std::vector<ImageConfigParams> image_config_params;
   std::function<std::unique_ptr<FlowProgrammer>()> create_flow_programmer;
   std::function<std::unique_ptr<TrafficHelper>()> create_traffic_helper;
@@ -43,8 +43,7 @@ struct NsfTestParams {
   std::function<std::vector<std::unique_ptr<ComponentValidator>>()>
       create_component_validators;
   std::function<std::unique_ptr<thinkit::SSHClient>()> create_ssh_client;
-  // TODO: This is a temporary workaround and
-  // needs to be removed once the testbeds are ready.
+  std::function<std::vector<std::string>(NsfUpgradeScenario)> get_test_case_ids;
   bool enable_interface_validation_during_nsf = true;
 };
